project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "minsearch>=0.0.5",
    "openai>=1.109.1",
    "pydantic-ai>=1.0.10",
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
